+++ modules/Material/ApplicationWindow.qml
@@ -102,42 +102,13 @@ Controls.ApplicationWindow {
         onReplaced: __toolbar.replace(page)
     }
 
+    OverlayLayer {
+        id: dialogOverlayLayer
+        objectName: "dialogOverlayLayer"
+    }
 
-    Rectangle {
+    OverlayLayer {
         id: overlayLayer
-        objectName: "overlayLayer"
-
-        anchors.fill: parent
-        z: 100
-
-        property Item currentOverlay
-        color: "transparent"
-
-        states: State {
-            name: "ShowState"
-            when: overlayLayer.currentOverlay != null
-
-            PropertyChanges {
-                target: overlayLayer
-                color: currentOverlay.backdropColor
-            }
-        }
-
-        transitions: Transition {
-            ColorAnimation {
-                duration: 300
-                easing.type: Easing.InOutQuad
-            }
-
-        }
-
-        MouseArea {
-            anchors.fill: parent
-            enabled: overlayLayer.currentOverlay != null
-            hoverEnabled: enabled
-            onClicked: overlayLayer.currentOverlay.close()
-            onWheel: wheel.accepted = true
-        }
     }
 
     width: units.dp(800)