+++ modules/Material/Dialog.qml
@@ -28,6 +28,8 @@ Popover {
 
     anchor: Item.Center
 
+    overlayLayer: "dialogOverlayLayer"
+
     backdropColor: Qt.rgba(0, 0, 0, 0.3)
 
     width: Math.max(dialogContainer.width, minimumWidth)